PR : Feat/knowledge base infra(#64)

+++ src/main/java/com/imyme/mine/domain/card/dto/CardDetailResponse.java
@@ -13,8 +13,8 @@ public record CardDetailResponse(
     Long keywordId,
     String keywordName,
     String title,
-    Integer bestLevel,
-    Integer attemptCount,
+    Short bestLevel,
+    Short attemptCount,
     LocalDateTime createdAt,
     LocalDateTime updatedAt,
     List<AttemptResponse> attempts
