+++ src/main/java/com/imyme/mine/domain/card/dto/CardListItem.java
@@ -10,8 +10,8 @@ public record CardListItem(
     Long keywordId,
     String keywordName,
     String title,
-    Integer bestLevel,
-    Integer attemptCount,
+    Short bestLevel,
+    Short attemptCount,
     LocalDateTime createdAt
 ) {
     /**
